backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-vision
google-genai
python-dotenv==1.0.0
